# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MobileCRM API Documentation",
        "description": "MobileCRM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/business/hotel/room-types": {
            "get": {
                "tags": [
                    "Hotel - Room Types"
                ],
                "summary": "List Hotel Room Types",
                "description": "Returns all room categories (e.g. Deluxe AC, Suite) for the active business.",
                "operationId": "6f3ea55dbe0109e88c7c733031680267",
                "responses": {
                    "200": {
                        "description": "Room types retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hotel - Room Types"
                ],
                "summary": "Create Room Type",
                "description": "Creates a new room category with pricing and amenities.",
                "operationId": "4b6a8d413ebbfbf1c649adb3eefb30b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "base_price_weekday",
                                    "max_occupancy"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Deluxe AC"
                                    },
                                    "short_code": {
                                        "type": "string",
                                        "example": "DLX"
                                    },
                                    "base_price_weekday": {
                                        "type": "number",
                                        "example": 2000
                                    },
                                    "base_price_weekend": {
                                        "type": "number",
                                        "example": 2500
                                    },
                                    "base_price_peak": {
                                        "type": "number",
                                        "example": 3500
                                    },
                                    "max_occupancy": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room type created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/business/hotel/room-types/{id}": {
            "get": {
                "tags": [
                    "Hotel - Room Types"
                ],
                "summary": "Get Room Type",
                "operationId": "5f168d3a74d1b66de8f5c4d4283bc498",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hotel - Room Types"
                ],
                "summary": "Update Room Type",
                "operationId": "2c7c142e2183e9b2f3aaa54370aba04a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hotel - Room Types"
                ],
                "summary": "Delete Room Type",
                "operationId": "1290d59698741759e439f9236884f064",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "422": {
                        "description": "Cannot delete — rooms exist"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer Token to authenticate",
                "scheme": "bearer"
            }
        }
    },
...
}